+++ src/threadx/SennaTech/Core/Src/sensor/ultrasonic/front_ultrasonic.c
@@ -1,0 +1,30 @@
+#include "srf08.h"
+
+VOID read_front_ultrasonic_distance_cm(int32_t *front_distance_cm)
+{
+    // ~1 second between measurements to avoid flooding queues traffic
+    tx_thread_sleep(ULTRASONIC_PERIOD_TICKS * 10);  
+    // Trigger a new ranging cycle
+    if (srf08_trigger(&hi2c2, SRF08_DEFAULT_ADDR) != TX_SUCCESS)
+    {
+        uart_send("[SRF08] Trigger failed\r\n");
+        tx_thread_sleep(ULTRASONIC_PERIOD_TICKS);
+        *front_distance_cm = -1;
+        return ;
+    }
+
+    //  Wait for the sensor to finish ranging (~65ms minimum)
+    tx_thread_sleep(SRF08_RANGING_DELAY_MS * TX_TIMER_TICKS_PER_SECOND / 1000);
+
+    // Read the result
+    if (srf08_read_cm(&hi2c2, SRF08_DEFAULT_ADDR, front_distance_cm) != TX_SUCCESS)
+    {
+        uart_send("[SRF08] Read failed\r\n");
+        tx_thread_sleep(ULTRASONIC_PERIOD_TICKS);
+        *front_distance_cm = -1;
+        return ;
+    }
+    uart_send("Distance (cm): ");
+    uart_send_int(*front_distance_cm);
+    uart_send("\r\n");
+}
